Removed the recipe for the Brick Mold
+comments

I also tried to add the recipe for the bricks to the BWM kiln but as it turns out the BWM kiln doesn't support recipes with items (only blocks) so yeah that didn't work and it just broke the whole script. And I don't think I will be adding the PrimalTerra anytime soon until it will have proper configuration options since I don't want to have Hibachis with Inventories.